[Framework] Routing algorithm avoid `Location`s
The routing algorithm of the `Map` is now able to avoid `Location`s.
To also accept `Location`s, the type of the `avoid` parameter was
changed to `Collection<Position>`.

Because of this, you may need to change types at certain locations of
your code. Apart from this, your code should not be affected by this
change.

Additionally, the `Validator` now checks for collisions at `Location`s.