[Framework] Add Gradle task for submission
A new Gradle task "submissionZip" was added that archives any PDF
present in the project root directory as well as the "src" directory
into a single zip file. The file is called "submission.zip".

Since this was the name of the zip file created in "srcZip", "srcZip"
now creates an archive called "sources.zip".